Improve debug functionality

+++ trajectory/body.h
@@ -140,7 +140,8 @@ void update_state_vectors(Body *sys, uint64_t count, Time dt);
 /// Updates every body in a system at time dt. Uses a constant timestep for all bodies, which is not optimal.
 void system_update(Body *sys, uint64_t count, Time dt, Time *t);
 
-uint64_t system_total_energy(Body *sys, uint64_t count);
+/// Returns the total energy of the system, in terajoules (1e12*J)
+long double system_total_energy(Body *sys, uint64_t count);
 
 void print_body_info(Body b);
